Refacroring: rename lib/metacom.js to lib/client.js
