+++ metacom.js
@@ -1,4 +1,4 @@
 'use strict';
 
-module.exports = require('./lib/metacom.js');
+module.exports = require('./lib/client.js');
 module.exports.Server = require('./lib/server.js');
